requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
pytest==7.4.3
allure-pytest==2.13.2
pytest-xdist==3.5.0
